timer: change to use time_mono (api break!)
Remove timer_add() in favor of explicit timer_addrel and timer_addmono.

Someone hit a real-life case where time went backwards, and we asserted.
The correct fix is to use time_mono() where available, but as all known
users actually want a relative timeout, have a helper for that case.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>