.travis.yml: Remove external libraries
The new makefile excludes all modules with external library dependencies
from the usual make check, meaning the libraries are not needed for the
Travis build.

Although it would be nice to include these modules, there are a bunch of
problems with sorting out a correct environment for them.  For now, just
leave them out.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>