time: Change TIME_HAVE_MONOTONIC to be 0/1 rather than defined/undefined
Generally, ccan config variables are always defined as either 0 or 1 and
are tested with #if.  That's instead of being being either defined or
undefined and tested with #ifdef.

TIME_HAVE_MONOTONIC breaks that convention.  This can cause warnings in
ccan/timer which uses it assuming the 0/1 convention.  Change it to remove
that warning.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>