read_write_all: fix read_all() to always set errno on failure
read(2) will return 0 to designate end-of-file and in this case it will
not set errno, but read_all() will return false, leaving errno as it
were before the call.

An EOF is not an error for read(2) itself, but read_all() is supposed to
read exactly the requested number of bytes (not less), so set errno to
EBADMSG "Bad message" in this case.