minmax: don't nest min/max in clamp to avoid -Wshadow warnings
Using clamp from minmax.h with -Wshadow (GCC 11 and 8, haven't tested others)
generates a bunch of warnings because min and max both declare _a, _b vars.
Store the result of min to a temporary variable instead of nesting the calls
this way min's variables are not in scope of max.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>