bitmap: Avoid shadowing type name with parameter name
'bitmap' is a typedef name in this module, and we also use it as a
parameter name in a bunch of places.  Although that's possible, because
types and variables live in different namespaces, it's probably not that
wise.  It also appears to trip warnings with at least some options and
compiler versions.

Reported-by: Ashok Raj <ashok.raj@intel.com>

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>