ccan/io: fix occasional immediate close after io_shutdown()
If we're doing a duplex connection, we could be still trying to write:
that will fail (since we called shutdown on the write side), and we'll
respond by closing the socket.

The reason we're doing shutdown() is to ensure the final bytes are
written: doing a close() breaks this, and so the result is lost final
bytes.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>